# crates/server/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Terrarium API",
    "version": "1.0.0",
    "description": "Deploy programmable creatures into a persistent real-time simulation.\n\nAuthenticate with a Firebase ID token or an API key (`tr_…`) in the `Authorization: Bearer` header.\n\nMint API keys via `POST /v1/api-keys` while signed in."
  },
  "servers": [
    { "url": "/api", "description": "Production (Firebase Hosting proxy)" },
    { "url": "", "description": "Local dev (direct to :8080)" }
  ],
  "tags": [
    { "name": "public", "description": "No authentication required" },
    { "name": "account", "description": "Account and API keys" },
    { "name": "world", "description": "Simulation state" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Firebase ID token or Terrarium API key (`tr_…`)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "ok" },
          "tick_hz": { "type": "integer", "example": 10 }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "uid": { "type": "string" },
          "credits": { "type": "integer", "format": "int64" }
        }
      },
      "Creature": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "x": { "type": "integer" },
          "y": { "type": "integer" },
          "energy": { "type": "integer", "format": "int64" },
          "health": { "type": "integer" },
          "max_health": { "type": "integer" },
          "owner_uid": { "type": "string" },
          "program_hash": { "type": "string" }
        }
      },
      "Tile": {
        "type": "object",
        "properties": {
          "x": { "type": "integer" },
          "y": { "type": "integer" },
          "kind": { "type": "integer", "description": "1=solid, 3=corpse, 4=food" },
          "energy": { "type": "integer", "format": "int64" },
          "death_reason": { "type": "string" }
        }
      },
      "WorldResponse": {
        "type": "object",
        "properties": {
          "deploy_cost": { "type": "integer", "format": "int64" },
          "corpse_energy": { "type": "integer", "format": "int64" },
          "creatures": { "type": "array", "items": { "$ref": "#/components/schemas/Creature" } },
          "tiles": { "type": "array", "items": { "$ref": "#/components/schemas/Tile" } }
        }
      },
      "DeployRequest": {
        "type": "object",
        "required": ["x", "y", "code", "energy"],
        "properties": {
          "x": { "type": "integer" },
          "y": { "type": "integer" },
          "code": { "type": "string", "description": "WAT source or label when using wasm_b64" },
          "energy": { "type": "integer", "format": "int64", "description": "Extra glims beyond corpse floor" },
          "wasm_b64": { "type": "string", "description": "Precompiled WASM (base64, max 64 KB)" }
        }
      },
      "DeployResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "x": { "type": "integer" },
          "y": { "type": "integer" },
          "energy": { "type": "integer", "format": "int64" },
          "credits": { "type": "integer", "format": "int64" }
        }
      },
      "MintApiKeyRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "ci-deploy" }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "prefix": { "type": "string", "description": "First chars of the key for identification" },
          "created_at": { "type": "string" },
          "last_used_at": { "type": "string" }
        }
      },
      "MintApiKeyResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiKey" },
          {
            "type": "object",
            "properties": {
              "secret": {
                "type": "string",
                "description": "Full API key — returned once at creation"
              }
            },
            "required": ["secret"]
          }
        ]
      },
      "ApiKeyList": {
        "type": "object",
        "properties": {
          "keys": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKey" } }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["public"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Server is up",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/world": {
      "get": {
        "tags": ["world", "public"],
        "summary": "World state",
        "description": "Current creatures, tiles, and deploy costs (REST). Live updates use WebSocket deltas.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorldResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/world/ws": {
      "get": {
        "tags": ["world", "public"],
        "summary": "World WebSocket",
        "description": "Real-time deltas. Upgrade to WebSocket; messages are JSON deltas (`full: true` on connect/resync).",
        "responses": {
          "101": { "description": "Switching Protocols" }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": ["account"],
        "summary": "Current account",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/api-keys": {
      "get": {
        "tags": ["account"],
        "summary": "List API keys",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyList" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["account"],
        "summary": "Mint API key",
        "description": "Creates a new key. The full `secret` is returned once — store it safely.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MintApiKeyRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MintApiKeyResponse" }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "delete": {
        "tags": ["account"],
        "summary": "Revoke API key",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Revoked" },
          "404": { "description": "Not found" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/deploy": {
      "post": {
        "tags": ["world"],
        "summary": "Deploy creature",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeployRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeployResponse" }
              }
            }
          },
          "402": { "description": "Insufficient credits" },
          "409": { "description": "Cell occupied" },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}